update CQP rules and test files
