+++ test/rules/cqp_dx/cqp.log
@@ -3,5 +3,5 @@
  40CW  21-Nov-23 22:35 0003  AB1CCC         ---  ---  123 KING TULA KING TUL 3         
  40CW  21-Nov-23 22:36 0004  AB1DDD         ---  ---  27TULA                 3         
  40CW  21-Nov-23 22:36 0005  AB1EEE         ---  ---  64ORAN        ORAN     3         
- 40SSB 21-Nov-23 22:37 0006  AB1FFF         ---  ---  8  BUTT       BUTT     2         
- 40SSB 21-Nov-23 22:38 0007  AB1GGG         ---  ---  5 TULA KERN   KERN     2         
+ 40SSB 21-Nov-23 22:37 0006  AB1FFF         ---  ---  8  BUTT       BUTT     3         
+ 40SSB 21-Nov-23 22:38 0007  AB1GGG         ---  ---  5 TULA KERN   KERN     3
